% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_obtanin_file_path.R
\name{fn_obtanin_file_path}
\alias{fn_obtanin_file_path}
\title{Buscar path de módulo de forma recursiva}
\usage{
fn_obtanin_file_path(root_path, file_name)
}
\arguments{
\item{root_path}{Carpeta base donde empezar a buscar (ej. local_folder_path)}

\item{file_name}{Nombre exacto del archivo .R}
}
\description{
Buscar path de módulo de forma recursiva
}
